# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks in a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: assigneeId
          in: query
          required: false
          schema:
            type: integer
        - name: labelId
          in: query
          required: false
          schema:
            type: integer
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high]
        - name: due
          in: query
          required: false
          schema:
            type: string
            enum: [overdue, next7, none]
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task in a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Deleted
  /tasks/{taskId}/move:
    post:
      operationId: moveTask
      tags: [tasks]
      summary: Move a task to a column and position
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskMove"
      responses:
        "200":
          description: Moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /tasks/{taskId}/activity:
    get:
      operationId: listTaskActivity
      tags: [activity]
      summary: List activity history for a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Label:
      type: object
      required: [id, projectId, name, color, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
...
    Task:
      type: object
      required: [id, projectId, columnId, title, type, priority, position, completed, checklistTotal, checklistDone, createdAt, labels, videoLinks]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        columnId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [standard, video]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
        position:
          type: integer
        mindmapId:
          type: ["integer", "null"]
        assigneeId:
          type: ["integer", "null"]
        assignee:
          type: ["object", "null"]
          required: [id, name]
          properties:
            id:
              type: integer
            name:
              type: string
            avatarUrl:
              type: ["string", "null"]
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
        checklistTotal:
          type: integer
        checklistDone:
          type: integer
        createdAt:
          type: string
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        videoLinks:
          type: array
          items:
            $ref: "#/components/schemas/VideoLink"
    TaskInput:
      type: object
      required: [title, columnId]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        type:
          type: string
          enum: [standard, video]
        columnId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
        assigneeId:
          type: ["integer", "null"]
        labelIds:
          type: array
          items:
            type: integer
        insertAt:
          type: string
          enum: [start, end]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [standard, video]
        columnId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
        position:
          type: integer
        mindmapId:
          type: ["integer", "null"]
        assigneeId:
          type: ["integer", "null"]
        completed:
          type: boolean
        labelIds:
          type: array
          items:
            type: integer
    VideoLink:
      type: object
      required: [id, taskId, url, position, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        url:
          type: string
        label:
          type: ["string", "null"]
        position:
          type: integer
        createdAt:
          type: string
...
    TaskMove:
      type: object
      required: [columnId, position]
      properties:
        columnId:
          type: integer
        position:
          type: integer
    UserMini:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
...
    Activity:
      type: object
      required: [id, taskId, action, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        action:
          type: string
          enum: [created, moved, completed, reopened, assignee_changed, due_changed, timer_started, timer_paused, timer_finished]
        detail:
          type: ["string", "null"]
        actor:
          oneOf:
            - $ref: "#/components/schemas/UserMini"
            - type: "null"
        createdAt:
          type: string
...